Add files via upload
Two examples are given: one cough record in mp3 format and its corresponding spectrogram as the CNN input.

In this dataset, the length of each record is not the same. The original sampling rate is 48 kHz.
The spectrogram size is fixed to 129*199. The x-axis represents time, fixed to 10 seconds (zero-padding). The y-axis represents frequency bins (256 points FFT after downsampling to 24 kHz).